Move readonly Onyx aliases to types/utils/ReadonlyOnyx.ts, drop ReadonlyOnyx for ReadonlyDeep
