+++ src/libs/actions/IOU/BulkEdit.ts
@@ -46,7 +46,7 @@ import type {SearchResultDataType} from '@src/types/onyx/SearchResults';
 import type {TransactionChanges} from '@src/types/onyx/Transaction';
 
 import type {NullishDeep, OnyxCollection, OnyxEntry, OnyxUpdate} from 'react-native-onyx';
-import type {ValueOf} from 'type-fest';
+import type {ReadonlyDeep, ValueOf} from 'type-fest';
 
 import {deepEqual} from 'fast-equals';
 // lodashUnionBy de-dupes recent attendees by email/displayName in one pass; no lodash-free equivalent is used here
@@ -62,7 +62,7 @@ import {getUpdatedMoneyRequestReportData} from './MoneyRequestBuilder';
  * while each value keeps the readonly shape it has when it comes out of an Onyx write input (`OnyxUpdate['value']`).
  */
 type SnapshotDataDraft = {
-    -readonly [TKey in keyof NullishDeep<SearchResultDataType>]: OnyxTypes.ReadonlyOnyx<NullishDeep<SearchResultDataType>[TKey]>;
+    -readonly [TKey in keyof NullishDeep<SearchResultDataType>]: ReadonlyDeep<NullishDeep<SearchResultDataType>[TKey]>;
 };
 
 type BulkEditWriteOnyxData = {
