+++ src/types/onyx/index.ts
@@ -196,7 +196,6 @@ import type SuggestedAgentRule from './SuggestedAgentRule';
 import type SupportalPermissionDenied from './SupportalPermissionDenied';
 import type Task from './Task';
 import type Transaction from './Transaction';
-import type {ReadonlyOnyx, ReadonlyOnyxCollection, ReadonlyOnyxEntry, ReadonlyOnyxInputOrEntry} from './Transaction';
 import type TransactionPending3DSReview from './TransactionPending3DSReview';
 import type TransactionsPending3DSReview from './TransactionsPending3DSReview';
 import type {TransactionViolation, ViolationName} from './TransactionViolation';
@@ -361,10 +360,6 @@ export type {
     TaxRates,
     TaxRatesWithDefault,
     Transaction,
-    ReadonlyOnyx,
-    ReadonlyOnyxEntry,
-    ReadonlyOnyxInputOrEntry,
-    ReadonlyOnyxCollection,
     MerchantRuleSuggestion,
     MergeTransaction,
     TransactionViolation,
